Support for DELAY as a floating point number including less than 1 second while maintaining backward compatibiitiy with integer arguments.

+++ image.h
@@ -26,7 +26,7 @@
 
 typedef struct {
 	Imlib_Image im;
-	unsigned int delay;
+	float delay;
 } img_frame_t;
 
 typedef struct {
